Refactor public data package
